requirements.txt
Flask
Flask-SocketIO
Flask-Session
Werkzeug
PyCryptodome
python-dotenv
eventlet
Flask-SQLAlchemy
SQLAlchemy
